Test ping b/f ssh for JHs, incr. wait time to 70s.
We've had lots of ALARMS after testjhinet which did not result in
a database entry, as the JumpHost creation was just slow, but not
failing and wait222 then actually waited long enough.
Let's reduce these here.
Also: Output real timeout (70s now, was only 28s before and not 4x25s).