v1.48: Fix reshuffling of 2ndary ports
One of the issues with reshuffling the 2ndary ports was that we set
IP addresses manually and we need to set them right ...
For this we need a stable ordering when we discover the env with
collectRes. Refactor this into orderVMs().

Inc config2ndNICs, let's issue commands to clear ip rules and
old IPs from eth1, so we get a consistent state. Note: There might
be an easier way -- as we have no ability to actually disable DHCP, we
could as well rely on it :-(

In calcRedirs(), avoid calling neutron port-show N times but rather
parser the port-list output N times.

Fix -4 (reshuffling) in DEPLOY situation. (This was disabled due to a
broken condition test.)