1.54: Support logging additional info after wait error.
When we use wait[List]Resources, we consider "error" a final state and stop
waiting for this particular resource.
This is reflected by the exit code.
New: Collect these failed resources in a global array (ERRRSC[*]) and provide
a helper handleWaitErr that does an operation on each of them (typically
XXX show) to collect more diagnostics for the log file.