Wait for LB to be gone, so nets can be deleted.
Occasionally, removing a LB takes very long, causing the JH subnet and
network deletion to fail, as the LB still has ports open on thoese nets.
This leads to false alarms. (There is an extra cleanup step later that
already took care of deleting the subnets/networks anyway, so we did not
leave orphans around.)

Fix this by waiting for the LB to be gone in deleteLBs.
This will slowdown the iterations a bit ...

Signed-off-by: Kurt Garloff <kurt@garloff.de>