Partially created nets/subnets might lack tag.
So we have observed timeouts in network creation (52s!).
The networks eventually still appear, however, sometimes without
the set tag. (Is the tag setting actually what causes the timeout?)
This causes trouble, as we filter for set tags in our cleanup routines.
So do some cleanup for nets/subnets without actually filtering
for the tag, so we catch them anyway.

Signed-off-by: Kurt Garloff <kurt@garloff.de>