Add description to floating IPs.
The tagging is not as reliable as we'd need it to be :-(
So put a description to floating IPs in addition.
This could be used to automatically clean them up from time to time.
(This cleanup is not yet implemented.)

Note: There were a few neutron floating-ip-XXX commands in the source
file (with XXX = create, list, delete). These won't work, as they need
to be written floatingip-XXX. This was harmless thus far, as we
typically translate these commands into openstack CLI commands (where
they become openstack floating ip XXX) and the translator was tolerant
against the extra dash. This was corrected.

Signed-off-by: Kurt Garloff <kurt@garloff.de>