Avoid collecting stats for 2nd waitLB.
Before delete the LBs, we give them a second chance to appear,
so we increase the chances for successful cleanup. In the best case,
it's a noop (causing us 2s longer cycle time). In the worst case,
we'll wait for another couple of minutes with no success.

This however had a side effect of confusing our stats collection
on LB creation times, as we measured a second time. This is removed
now.

Signed-off-by: Kurt Garloff <kurt@garloff.de>