1.80: Better iperf wait, retry. Report rel. iter time.
iperf3 may take longer to appear if repo servers are slow.
So wait longer for cloud-init to do its job.
Even if completed on the JumpHost, the clients may need a bit more
time, so give it an additional 30s and retry if iperf3 failed.

The total time per iteration is an important number; send it
to telegraf. We send 10x the relative time (vs. the calculated
max cycle time), so it fits well into the benchmark panel.

Signed-off-by: Kurt Garloff <kurt@garloff.de>