Add log_grafana wrapper. ssh errors to logfile.
We had tons of if test "$GRAFANA"; ... curl ...; fi in the code.
Put this into the log_grafana wrapper, also optionally call rc2grafana there.
We drop the grafana.log file -- it has no valuable content.

When using ssh in testlsandping, let's ensure that errors are visible in
the logfile. Also call the ping command with a timeout 24, so we don't
wait ages for DNS. (ping itself is bounded to 10s normally.)

When enabling port forwarding, we also ssh to the JH, record errors
there as well. (Both console and logfile were missing here previously.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>