Retry removing LB after waitdel.
When JH creation fails, we clean up things in reverse order.
The LB is then typcally still in PENDING_CREATE, which means
we can't delete it yet. Waiting it to vanish then won't work
either.
If that happens, we should try to delete it again and wait
again.

This addresses the issue observed in #96.

Signed-off-by: Kurt Garloff <scs@garloff.de>