Wait shorter b/f retrying LB delete.
Also for the implementation to work, we need to unset LBDSTATS
before deleting again (otherwise the array grows larger than the
list of resources, causing the wait loop to not break once all
resources are gone).
When retrying, we wait 5s rather than 2s now, increasing the chances
that re-deleting actually works.

This fixes #96.

Signed-off-by: Kurt Garloff <scs@garloff.de>