Handle neutron not reporting device_id.
We currently match ports by looking for the VM UUID in the port's
device_id. This does no longer seem to work reliably on PCO; the field
is not returned by the port list command.
So look for the IP address reported from server list and match it
with the ports to determine port ID. Sort of ugly, but there is no
straightforward way to get port <-> VM relationships without looking
at port details or matching both subnet and IP address.

Signed-off-by: Kurt Garloff <kurt@garloff.de>