Fix/assign extnet earlier (#107)
* Handle neutron not reporting device_id.

We currently match ports by looking for the VM UUID in the port's
device_id. This does no longer seem to work reliably on PCO; the field
is not returned by the port list command.
So look for the IP address reported from server list and match it
with the ports to determine port ID. Sort of ugly, but there is no
straightforward way to get port <-> VM relationships without looking
at port details or matching both subnet and IP address.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Simplify old code path.

We already know that we only go here if we're using the legacy tooling.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Assign external net to router before creating VMs.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Revert "Assign external net to router before creating VMs."

This reverts commit b0a1991f38c4bf09499c07a4d6c4d40a3d558b00.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Assign external net to router before creating VMs.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Clean up router when external-net-list fails.

Also, ignore when setting external gateway fails.
(The reason is that this is not required on OTC and the error is
thus harmless.)

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Bump version number to 1.85.

The change (assigning the external net early) is significant enough
that we want to see from the version number whether or not we have it.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

Signed-off-by: Kurt Garloff <kurt@garloff.de>