LB member subnet is JHSUBNET[0].
Explicitly setting the --subnet-id for amphorae to the backend
member actually breaks the LB's connection to the backend members.
So we set it to the subnet from the VIP; as we access the LB via
the floating IP, the request does indeed originate from the LB's
VIP address in the VIP subnet (which we set to the JHSUBNET[0]).

When using ovn provider, this makes the LB work IF the requests
come from a host in the JHSUBNET, but not from the floating IP.
So currently, this breaks.

Sidenote: health-monitor is not supported for ovn provider pools.

Signed-off-by: Kurt Garloff <kurt@garloff.de>