Feat/lb provider flavor (#118)
* Support other LB providers (ovn) with -LP.

* Allow passing loadbalancer provider via -LP <PROVIDER>
* This results in a TCP loadbalancer with algo SOURCE_IP_PORT
  (which is the only algo that the ovn provider supports)
* When configuring members, we actually had ommitted the --subnet-id
  parameter previously, as it was unneeded and lead to a failure
  with token authentication.
  This needed changing, as the member subnets need to be set
  explicitly when using the OVN provider.

Still: It does not (yet?) work.
Maybe having the VIP (for the listener) and the member
ports in different networks is not supported?

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* LB member subnet is JHSUBNET[0].

Explicitly setting the --subnet-id for amphorae to the backend
member actually breaks the LB's connection to the backend members.
So we set it to the subnet from the VIP; as we access the LB via
the floating IP, the request does indeed originate from the LB's
VIP address in the VIP subnet (which we set to the JHSUBNET[0]).

When using ovn provider, this makes the LB work IF the requests
come from a host in the JHSUBNET, but not from the floating IP.
So currently, this breaks.

Sidenote: health-monitor is not supported for ovn provider pools.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* SG in VM needs to allow port 80 from 0/0 for OVN.

The requests come from the real client IP (which can be the internet,
i.e. 0/0) and not the LB's virtual IP. Thus adjust the security
group to allow for it.

The OVN provider does not support the health monitor, unfortunately.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Adjust max cycle time if we don't kill LB members.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Improve warning when LB backend kills are skipped.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Set the member's subnet to the backend server subnets.

This causes the health-monitor to detect members with error operating
status.
It does correctly take them out of the rotation when accessed via the
VIP, unfortunately not via the Floating IP though.
https://bugs.launchpad.net/neutron/+bug/1956035

The api_monitor.sh does thus report a few errors on each iteration.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

---------

Signed-off-by: Kurt Garloff <kurt@garloff.de>