Cleanup LBs when Octavia times out.
In gx-scs currently, the Loadbalancer service is extremely
slow. When createing the LB instance (with two amphorae),
we time out in the creation API call (which does NOT wait
for them to become active). So we end up testing without
the LB test. However, the LBs eventually come up and we fail
to expect this and fail to clean up. So more and more LBs
are hanging around until the 200 iterations are over.
We also fail to cleanup networks by consequence.

Change this: In cleanup, if LBs are enabled at all, but
there were errors, looks for LBs and try to delete them.
So we stop collecting them and also avoid failed network
and router cleanup.

Signed-off-by: Kurt Garloff <kurt@garloff.de>